backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
decorator==4.4.2
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
dynaconf==3.1.4
environs==9.3.1
hvac==0.10.8
idna==2.10
django-rest-knox==4.1.0
loguru==0.5.3
Markdown==3.3.4
marshmallow==3.10.0
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
validators==0.18.2
win32-setctime==1.0.3
drf-yasg==1.20.0
knox==0.1.11